findakey_game.py
Python game.

You move on a 10x10 board.
In a given game, you move around the board with the 'W/S/A/D' keys. You have to find the key, which is placed in a random place. Don't worry, you'll see if you're going in the right direction.

The program is a trial task for training purposes.